fastmcp-agents-library/agents/fastmcp-agents-library-agent-documentation-maintainer/pyproject.toml
[project]
name = "fastmcp-agents-library-agent-documentation-maintainer"
version = "0.5.7"
description = "An agent for maintaining documentation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp",
    "fastmcp-agents-library-mcp",
    "fastmcp-ai-agent-bridge-pydantic-ai>=0.1.0",
    "gitpython>=3.1.44",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.4.4",
]

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "fastmcp_agents.library.agent.documentation_maintainer"


[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
fastmcp-agents-library-mcp = { workspace = true }


[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "skip_on_ci: Skip running the test when running on CI",
]
addopts = "-s" 
norecursedirs = ["playground"]
[dependency-groups]
dev = [
    "ruff>=0.12.4",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
maintain-documentation = "fastmcp_agents.library.agent.documentation.maintainer:run"


[tool.ruff]
extend="../../../pyproject.toml"

[tool.pyright]
extends = "../../../pyproject.toml"
